+++ src/config.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-#define FIRMWARE_VERSION "0.1.2"
+#define FIRMWARE_VERSION "0.1.3"
 
 // WiFi AP
 #define AP_PASSWORD "moduleaircfg"